Always use /api/posts/{POST_ID} to avoid stale-post-slug logic
If the post used to have a post slug, it follows the post until it gets a new post slug. If the post was moved, we end up with a situation where the collection/slug combination is out of sync, and so the API sends back a 404 (FIXME: the Swift package silently fails here).