Embed text editor fields into ScrollView (#227)
* Embed text editor fields into ScrollView

* Bump version/build number and update change log

* Fix truncated post badges on status change

When going from a shorter label (e.g., ‘LOCAL’) to a longer one (e.g., ‘PUBLISHED’), the SwiftUI render loop would not expand the size of the post badge to fit the extra characters, so on change it would be truncated (e.g., read ‘PUBL…’). This change fixes the bug.